[Refactor] extract HLA connector to separate file, merge redundant classes, and remove sliding_window terminology
